[feat] alert dialog 위치 수정

+++ core-network/src/main/java/com/ddd/pollpoll/core/network/model/PostPostRequest.kt
@@ -5,7 +5,7 @@ import com.ddd.pollpoll.Vote
 data class PostPostRequest(
     val categoryId: Int,
     val contents: String,
-    val milliseconds: Int,
+    val milliseconds: Long,
     val multipleChoice: Boolean,
     val pollItems: List<PollItem>,
     val title: String
